Various fixes, glyphs module
check_range changed to be a function using generics. Some if let
statements were replaced with matches, and cons conversion to lisp
object was made nicer.

Additionally, types and functions handing glyphs were moved to a
new module.
